Started with policies

+++ app/controllers/institutions_controller.rb
@@ -4,23 +4,28 @@ class InstitutionsController < ApplicationController
 
   add_breadcrumb 'Institutions', :institutions_path
 
-  expose(:institutions) { current_user.institutions }
   expose(:institution, attributes: :institution_params)
 
   def index
+    @institutions = authorize_access(Institution, "cdp:list_institutions")
   end
 
   def new
+    @institution = current_user.institutions.new
     set_institution_tab :new
   end
 
   def show
+    @institutions = authorize_access(institution, "cdp:list_institutions")
+
     add_breadcrumb institution.name, institution
     add_breadcrumb 'Overview'
     set_institution_tab :overview
   end
 
   def edit
+    return unless authorize_access(institution, "cdp:edit_institution")
+
     add_breadcrumb institution.name, institution
     add_breadcrumb 'Settings'
     set_institution_tab :settings
@@ -39,6 +44,8 @@ class InstitutionsController < ApplicationController
   end
 
   def update
+    return unless authorize_access(institution, "cdp:edit_institution")
+
     respond_to do |format|
       if institution.update(institution_params)
         format.html { redirect_to institution, notice: 'Institution was successfully updated.' }
@@ -51,6 +58,8 @@ class InstitutionsController < ApplicationController
   end
 
   def destroy
+    return unless authorize_access(institution, "cdp:delete_institution")
+
     institution.destroy
     respond_to do |format|
       format.html { redirect_to institutions_url }
